+++ src/pools/cluster/dynamic.ts
@@ -40,11 +40,6 @@ export class DynamicClusterPool<
   }
 
   /** @inheritDoc */
-  protected get maxSize (): number {
-    return this.max
-  }
-
-  /** @inheritDoc */
   protected get busy (): boolean {
     return this.full && this.internalBusy()
   }