+++ src/pools/cluster/fixed.ts
@@ -117,16 +117,6 @@ export class FixedClusterPool<
   }
 
   /** @inheritDoc */
-  protected get minSize (): number {
-    return this.numberOfWorkers
-  }
-
-  /** @inheritDoc */
-  protected get maxSize (): number {
-    return this.numberOfWorkers
-  }
-
-  /** @inheritDoc */
   protected get busy (): boolean {
     return this.internalBusy()
   }