+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ensure pool event `backPressure` is emitted.
 - Ensure pool event `full` is emitted only once.
+- Ensure worker node cannot be instantiated without proper arguments.
 
 ## [2.6.29] - 2023-08-18